BDCC/proj/1st/app/requirements.txt
google-cloud-bigquery
google-cloud-storage
Flask
tensorflow==2.8.0
Pillow==9.1.0
google-cloud-vision==2.7.2
gunicorn==19.9.0